Fixed type conversions in new API subscription timeout functionality

+++ src/api/cpp/GravitySubscriptionManager.h
@@ -57,8 +57,8 @@ private:
 	{
 		GravitySubscriptionMonitor* monitor;
 		int timeout;
-		long endTime;
-		long lastReceived;
+		int endTime;
+		int lastReceived;
 	} TimeoutMonitor;
 
     typedef struct SubscriptionDetails
